+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Skip step in profile now possible (name inside a profile step starts with "pause")
 - Library updates/Flutter version update
 - fixing interpolation issues during puring
+- Transport mode added to maintenance menu. Purging water out of de1 system.
 
 ### Visualizer changes
 You need to go to Settings/Cloud and Network/Visualizer and click on the *Login to Visualizer* button to get redirected to the Visualizer page.